libc: fix 56 orphaned class declarations + fix uuL/uL suffix corruption in hex literals
- Add missing 'public data class ClassName(' headers to 29 files with
  orphaned val blocks (generated code had stripped class declaration lines)
- Fix uuL -> uL suffix corruption introduced by type-mismatch fix attempt
- Fix 0x0000u_0001 -> 0x00000001u (u inserted mid-hex-literal) in 77 lines
- Fix 0x0000uL_0001 -> 0x00000001uL (uL inserted mid-hex-literal) in 29 lines

Build errors reduced from 5241 to 3363.

+++ src/commonMain/kotlin/io/github/kotlinmania/libc/unix/solarish/Mod.kt
@@ -601,9 +601,9 @@ public const val SS_DISABLE: CInt = 2
 public const val FIOCLEX: CInt = 0x20006601
 public const val FIONCLEX: CInt = 0x20006602
 public const val FIONREAD: CInt = 0x4004667f
-public const val FIONBIO: CInt = 0x8004667e
-public const val FIOASYNC: CInt = 0x8004667d
-public const val FIOSETOWN: CInt = 0x8004667c
+public const val FIONBIO: CInt = 0x8004667e.toInt()
+public const val FIOASYNC: CInt = 0x8004667d.toInt()
+public const val FIOSETOWN: CInt = 0x8004667c.toInt()
 public const val FIOGETOWN: CInt = 0x4004667b
 public const val SIGCHLD: CInt = 18
 public const val SIGCLD: CInt = SIGCHLD
@@ -772,7 +772,7 @@ public const val WTRAPPED: CInt = 0x02
 public const val WSTOPPED: CInt = WUNTRACED
 public const val WCONTINUED: CInt = 0x08
 public const val WNOWAIT: CInt = 0x80
-public const val AT_FDCWD: CInt = 0xffd19553
+public const val AT_FDCWD: CInt = 0xffd19553.toInt()
 public const val AT_SYMLINK_NOFOLLOW: CInt = 0x1000
 public const val AT_SYMLINK_FOLLOW: CInt = 0x2000
 public const val AT_REMOVEDIR: CInt = 0x1
@@ -1218,7 +1218,7 @@ public const val IFF_NOFAILOVER: CInt = 0x0008000000
 public const val IFF_FAILED: CInt = 0x0010000000
 public const val IFF_STANDBY: CInt = 0x0020000000
 public const val IFF_INACTIVE: CInt = 0x0040000000
-public const val IFF_OFFLINE: CInt = 0x0080000000
+public const val IFF_OFFLINE: CInt = 0x0080000000.toInt()
 public const val IFF_COS_ENABLED: CLongLong = 0x0200000000
 public const val IFF_PREFERRED: CLongLong = 0x0400000000
 public const val IFF_TEMPORARY: CLongLong = 0x0800000000
@@ -1435,7 +1435,7 @@ public val _COND_MAGIC: UShort = (0x4356).toUShort()
 public val _RWL_MAGIC: UShort = (0x5257).toUShort()
 public const val NCCS: ULong = 19uL
 public val LOG_CRON: CInt = 15 shl 3
-public val PTHREAD_MUTEX_INITIALIZER: PthreadMutexT = PthreadMutexT(pthreadMutexFlag1 = 0, pthreadMutexFlag2 = 0, pthreadMutexCeiling = 0, pthreadMutexType = PTHREAD_PROCESS_PRIVATE, pthreadMutexMagic = _MUTEX_MAGIC, pthreadMutexLock = 0, pthreadMutexData = 0)
+public val PTHREAD_MUTEX_INITIALIZER: PthreadMutexT = PthreadMutexT(pthreadMutexFlag1 = 0.toUShort(), pthreadMutexFlag2 = 0.toUByte(), pthreadMutexCeiling = 0.toUByte(), pthreadMutexType = PTHREAD_PROCESS_PRIVATE, pthreadMutexMagic = _MUTEX_MAGIC, pthreadMutexLock = 0uL, pthreadMutexData = 0uL)
 
 // PTHREAD_COND_INITIALIZER: PthreadCondT = pthread_cond_t { __pthread_cond_flag: [0 (initializer represented at the FFI boundary)
 public val PTHREAD_RWLOCK_INITIALIZER: PthreadRwlockT = PthreadRwlockT(pthreadRwlockReaders = 0, pthreadRwlockType = PTHREAD_PROCESS_PRIVATE, pthreadRwlockMagic = _RWL_MAGIC, pthreadRwlockMutex = PTHREAD_MUTEX_INITIALIZER, pthreadRwlockReadercv = PTHREAD_COND_INITIALIZER, pthreadRwlockWritercv = PTHREAD_COND_INITIALIZER)
